refactored common code getImageModel

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -111,13 +111,9 @@ public class ImageController {
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
     public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session) {
-        Image image = imageService.getImage(imageId);
 
         User loggeduser = (User) session.getAttribute("loggeduser");
-        String tags = convertTagsToString(image.getTags());
-        model.addAttribute("id", imageId);
-        model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
+        Image image = getImageModel(imageId, model);
         if (!checkImageOwner(image.getUser(), loggeduser)) {
             model.addAttribute("tags", image.getTags());
             String error = "Only the owner of the image can edit the image";
@@ -125,6 +121,21 @@ public class ImageController {
             return "/images/image";
         }
         return "images/edit";
+    }
+
+    /**
+     * Create a model by getting all details of Image from repository
+     * @param imageId
+     * @param model
+     * @return
+     */
+    private Image getImageModel(Integer imageId, Model model) {
+        Image image = imageService.getImage(imageId);
+        String tags = convertTagsToString(image.getTags());
+        model.addAttribute("id", imageId);
+        model.addAttribute("image", image);
+        model.addAttribute("tags", tags);
+        return image;
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -176,7 +187,16 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, Model model, HttpSession session) {
+        Image image = getImageModel(imageId, model);
+        User loggeduser = (User) session.getAttribute("loggeduser");
+        if (!checkImageOwner(image.getUser(), loggeduser)) {
+            model.addAttribute("tags", image.getTags());
+            String error = "Only the owner of the image can delete the image";
+            model.addAttribute("deleteError", error);
+            return "/images/image";
+        }
+
         imageService.deleteImage(imageId);
         return "redirect:/images";
     }
